[FLINK-12505][hive] Unify database operations to HiveCatalogBase from its subclasses
Currently each subclass of HiveCatalogBase has its own impl for catalog related database APIs and its own util class to support such impl. However, they share many common code. This PR unifies their common parts to make the code cleaner and more readable.

This closes #8433.
