+++ flink-connectors/flink-connector-hive/src/main/java/org/apache/flink/table/catalog/hive/HiveCatalogBase.java
@@ -20,6 +20,7 @@ package org.apache.flink.table.catalog.hive;
 
 import org.apache.flink.table.catalog.Catalog;
 import org.apache.flink.table.catalog.CatalogBaseTable;
+import org.apache.flink.table.catalog.CatalogDatabase;
 import org.apache.flink.table.catalog.ObjectPath;
 import org.apache.flink.table.catalog.exceptions.CatalogException;
 import org.apache.flink.table.catalog.exceptions.DatabaseAlreadyExistException;
@@ -128,6 +129,23 @@ public abstract class HiveCatalogBase implements Catalog {
 	 */
 	protected abstract Table createHiveTable(ObjectPath tablePath, CatalogBaseTable table);
 
+	/**
+	 * Create a CatalogDatabase from a Hive database.
+	 *
+	 * @param hiveDatabase a Hive database
+	 * @return a CatalogDatabase
+	 */
+	protected abstract CatalogDatabase createCatalogDatabase(Database hiveDatabase);
+
+	/**
+	 * Create a Hive database from a CatalogDatabase.
+	 *
+	 * @param databaseName name of the database
+	 * @param catalogDatabase a CatalogDatabase
+	 * @return a Hive database
+	 */
+	protected abstract Database createHiveDatabase(String databaseName, CatalogDatabase catalogDatabase);
+
 	@Override
 	public void open() throws CatalogException {
 		if (client == null) {
@@ -155,6 +173,19 @@ public abstract class HiveCatalogBase implements Catalog {
 	@Override
 	public String getDefaultDatabase() throws CatalogException {
 		return defaultDatabase;
+	}
+
+	@Override
+	public CatalogDatabase getDatabase(String databaseName)
+			throws DatabaseNotExistException, CatalogException {
+		try {
+			return createCatalogDatabase(client.getDatabase(databaseName));
+		} catch (NoSuchObjectException e) {
+			throw new DatabaseNotExistException(catalogName, databaseName);
+		} catch (TException e) {
+			throw new CatalogException(
+				String.format("Failed to get database %s from %s", databaseName, catalogName), e);
+		}
 	}
 
 	@Override
@@ -194,35 +225,26 @@ public abstract class HiveCatalogBase implements Catalog {
 		}
 	}
 
-	protected Database getHiveDatabase(String databaseName) throws DatabaseNotExistException {
-		try {
-			return client.getDatabase(databaseName);
-		} catch (NoSuchObjectException e) {
-			throw new DatabaseNotExistException(catalogName, databaseName);
-		} catch (TException e) {
-			throw new CatalogException(
-				String.format("Failed to get database %s from %s", databaseName, catalogName), e);
-		}
-	}
-
-	protected void createHiveDatabase(Database hiveDatabase, boolean ignoreIfExists)
+	@Override
+	public void createDatabase(String name, CatalogDatabase database, boolean ignoreIfExists)
 			throws DatabaseAlreadyExistException, CatalogException {
 		try {
-			client.createDatabase(hiveDatabase);
+			client.createDatabase(createHiveDatabase(name, database));
 		} catch (AlreadyExistsException e) {
 			if (!ignoreIfExists) {
-				throw new DatabaseAlreadyExistException(catalogName, hiveDatabase.getName());
+				throw new DatabaseAlreadyExistException(catalogName, name);
 			}
 		} catch (TException e) {
-			throw new CatalogException(String.format("Failed to create database %s", hiveDatabase.getName()), e);
+			throw new CatalogException(String.format("Failed to create database %s", name), e);
 		}
 	}
 
-	protected void alterHiveDatabase(String name, Database newHiveDatabase, boolean ignoreIfNotExists)
+	@Override
+	public void alterDatabase(String name, CatalogDatabase newDatabase, boolean ignoreIfNotExists)
 			throws DatabaseNotExistException, CatalogException {
 		try {
 			if (databaseExists(name)) {
-				client.alterDatabase(name, newHiveDatabase);
+				client.alterDatabase(name, createHiveDatabase(name, newDatabase));
 			} else if (!ignoreIfNotExists) {
 				throw new DatabaseNotExistException(catalogName, name);
 			}
